+++ .gitignore
@@ -2,3 +2,4 @@
 data/*
 dumps/*
 *.egg-info
+build/*
