+++ dashboards/tokenizer_review.py
@@ -4,7 +4,11 @@ import streamlit_pianoroll
 from fortepyan import MidiPiece
 from datasets import Dataset, load_dataset
 
-from object_generators.tokenizer_generator import tokenizer_info, name_to_factory_map, generate_tokenizer_with_streamlit
+from midi_tokenizers_generation.tokenizer_generator import (
+    tokenizer_info, 
+    name_to_factory_map, 
+    generate_tokenizer_with_streamlit,
+)
 
 
 @st.cache_data
