+++ midi_quantizers/relative_time_quantizer.py
@@ -2,7 +2,7 @@ import yaml
 import numpy as np
 import pandas as pd
 
-from quantizers.quantizer import MidiQuantizer
+from midi_quantizers.quantizer import MidiQuantizer
 
 
 class RelativeTimeQuantizer(MidiQuantizer):
@@ -31,7 +31,7 @@ class RelativeTimeQuantizer(MidiQuantizer):
         self._build_velocity_decoder()
 
     def _load_bin_edges(self):
-        artifacts_path = "artifacts/bin_edges.yaml"
+        artifacts_path = "midi_quantization_artifacts/bin_edges.yaml"
         with open(artifacts_path, "r") as f:
             bin_edges = yaml.safe_load(f)
