+++ midi_tokenizers_generation/base_tokenizer_generator.py
@@ -6,7 +6,7 @@ from midi_tokenizers.midi_tokenizer import MidiTokenizer
 from midi_tokenizers.no_loss_tokenizer import NoLossTokenizer
 from midi_tokenizers.one_time_tokenizer import OneTimeTokenizer
 from midi_tokenizers.quantized_midi_tokenizer import QuantizedMidiTokenizer
-from object_generators.quantizer_generator import name_to_quantizer_factory_map
+from midi_tokenizers_generation.quantizer_generator import name_to_quantizer_factory_map
 
 
 class TokenizerFactory:
