+++ midi_tokenizers_generation/tokenizer_generator.py
@@ -5,7 +5,7 @@ import streamlit as st
 
 from midi_tokenizers.midi_tokenizer import MidiTokenizer
 from midi_trainable_tokenizers.bpe_tokenizer import BpeMidiTokenizer
-from object_generators.base_tokenizer_generator import TokenizerFactory, name_to_base_factory_map
+from midi_tokenizers_generation.base_tokenizer_generator import TokenizerFactory, name_to_base_factory_map
 
 
 class BpeMidiTokenizerFactory(TokenizerFactory):
